fix: 修复 TrendingTheme/Swatch 因 Color(ULong) 解析 ARGB 导致启动崩溃
Compose 的 Color(value: ULong) 重载接收的是内部 packed 表示（高位含
color space 索引），并非 ARGB 整数；把 0xFF6750A4L.toULong() 传进去后，
getColorSpace-impl 读到的索引超出 RGB color spaces 数组长度，抛
ArrayIndexOutOfBoundsException: length=20; index=36，整个 App 启动即崩。

改用 Color(value: Long) 重载，按 ARGB int 解读。修复 TrendingTheme.kt
中 rememberDynamicMaterialThemeState 的 seedColor 入参，以及 SettingsScreen
中 ThemeSwatch 的色块填充色。

+++ shared/src/commonMain/kotlin/whl/trending/ai/ui/settings/SettingsScreen.kt
@@ -381,7 +381,7 @@ private fun ThemeSwatch(
     selected: Boolean,
     onClick: () -> Unit,
 ) {
-    val color = Color(seed.argb.toULong())
+    val color = Color(seed.argb)
     val name = stringResource(seed.nameRes)
     val border = if (selected) {
         BorderStroke(2.dp, MaterialTheme.colorScheme.outline)
